Re-organize ugly_widget
